Create functions in C++ and Java to create TileDB workspace
